pyproject.toml
[project]
name = "yomemoai-mcp"
version = "0.1.3"
description = "Model Context Protocol (MCP) server for YoMemoAI - enables AI assistants to save and retrieve encrypted memories"
readme = "README.md"
license = { text = "MIT" }
authors = [
    { name = "xxlv", email = "lvxiang119@gmail.com" }
]
requires-python = ">=3.12"
keywords = ["mcp", "mcp-server", "yomemoai", "memory", "ai", "claude", "cursor"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    "cryptography>=46.0.3",
    "fastmcp>=0.1.0",
    "mcp>=1.25.0",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
]

[project.urls]
Homepage = "https://github.com/yomemoai/python-yomemo-mcp"
Documentation = "https://github.com/yomemoai/python-yomemo-mcp#readme"
Repository = "https://github.com/yomemoai/python-yomemo-mcp"
Issues = "https://github.com/yomemoai/python-yomemo-mcp/issues"

[project.scripts]
yomemoai-mcp = "yomemoai_mcp.server:mcp.run"
memo-mcp = "yomemoai_mcp.server:mcp.run"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
